+++ tinper-bee/bee-timepicker/src/Timepicker.js
@@ -22,12 +22,12 @@ class Timepicker extends Component {
         //判断初始值是否合法，不合法则格式化，并触发onChange回调
         let value = this.props.value;
         if(value){
-            if(value.format){
+            if(typeof value == 'string'){
+                value = moment(`${moment().format('YYYY-MM-DD')} ${value}`);
+            }else if(value.format){
                 if(!value.isValid()){
                     value = moment(`${moment().format('YYYY-MM-DD')} ${value._i}`)
                 }
-            }else{
-                value = moment(`${moment().format('YYYY-MM-DD')} ${value}`);
             }
         }else{
             value = null
@@ -39,10 +39,12 @@ class Timepicker extends Component {
         if('value' in nextProps){
             let value = nextProps.value;
             if(value){
-                if(value.format){
-                    
-                }else{
-                    value = moment(`${moment().format('YYYY-MM-DD')} ${value}`)
+                if(typeof value == 'string'){
+                    value = moment(`${moment().format('YYYY-MM-DD')} ${value}`);
+                }else if(value.format){
+                    if(!value.isValid()){
+                        value = moment(`${moment().format('YYYY-MM-DD')} ${value._i}`)
+                    }
                 }
             }else{
                 value=null;