+++ tinper-bee/bee-datepicker/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.2.18"></a>
+## [2.2.18](https://github.com/tinper-bee/bee-datepicker/compare/v2.2.17...v2.2.18) (2019-11-14)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** form blur showTime  校验无法选择bug ([5c537a1](https://github.com/tinper-bee/bee-datepicker/commit/5c537a1))
+
+
+
 <a name="2.2.17"></a>
 ## [2.2.17](https://github.com/tinper-bee/bee-datepicker/compare/v2.2.16...v2.2.17) (2019-11-13)