+++ tinper-bee/bee-form-control/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.0.10"></a>
+## [2.0.10](https://github.com/tinper-bee/form-control/compare/v2.0.9...v2.0.10) (2019-11-14)
+
+
+### Features
+
+* **bee-form-control:** 新增 prefix suffix api ([037a6d0](https://github.com/tinper-bee/form-control/commit/037a6d0))
+
+
+
 <a name="2.0.9"></a>
 ## [2.0.9](https://github.com/tinper-bee/form-control/compare/v2.0.8...v2.0.9) (2019-09-24)