+++ tinper-bee/bee-input-number/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="2.2.1"></a>
+## [2.2.1](https://github.com/tinper-bee/bee-input-number/compare/v2.2.0...v2.2.1) (2019-11-14)
+
+
+
 <a name="2.2.0"></a>
 # [2.2.0](https://github.com/tinper-bee/bee-input-number/compare/v2.1.9...v2.2.0) (2019-11-13)