+++ tinper-bee/bee-timepicker/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="2.0.8"></a>
+## [2.0.8](https://github.com/tinper-bee/bee-timepicker/compare/v2.0.7...v2.0.8) (2019-11-14)
+
+
+
 <a name="2.0.7"></a>
 ## [2.0.7](https://github.com/tinper-bee/bee-timepicker/compare/v2.0.6...v2.0.7) (2019-10-29)